Move StatusReport to Screen
