+++ includes/classes/Screen.php
@@ -53,8 +53,11 @@ class Screen {
 
 		$this->sync_screen        = new Screen\Sync();
 		$this->health_info_screen = new Screen\HealthInfo();
+		$this->status_report      = new Screen\StatusReport();
+
 		$this->sync_screen->setup();
 		$this->health_info_screen->setup();
+		$this->status_report->setup();
 	}
 
 	/**
